Flask/requirements.txt
APScheduler==3.10.1
bcrypt==4.0.1
cachelib==0.9.0
click==8.1.3
Flask==2.2.3
Flask-Caching==2.0.2
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
psycopg2==2.9.5
PyJWT==2.6.0
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
SQLAlchemy==2.0.7
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
Werkzeug==2.2.3
